Created IBEP-20 v0.7.6
Compared to ERC-20, the BEP-20 is a fairly new standard, and due to that there are a few instances of clear code to be able to use by other developers. This is a mash up of the IERC20 and the an interface for BEP20, creating a mixture of a basic IBEP20 that can be used as a template for creating BEP20 tokens.
This interface syntax follows Solidity version 0.7.6, this is why the name of the file has a suffix of "_076".