Created thiefWallet.sol
This was an idea after checking the Wirex ecosystem, Wirex has hundreds if not thousands of contracts that you can send funds to and then make the contract send the funds again towards another wallet. It is not anonymous as Etherscan even is it has not been published and verified, Etherscan will show the ERC20 internal transaction towards the wallet in specific.